refactor(courses): give the collection picker its own hook
Three of Greptile's P1s landed in `use-course-card.ts`, which is a signal
about the file rather than three coincidences. Two were the same problem:
putting a course in a collection is not one write. The service only lets a
course join a collection its owner has saved, so the picker saves first,
and from there two writes fail independently in ways that need different
words and different recovery advice.

That consistency problem now lives in `useCollectionPicker`, away from
save and taken, which are each a single idempotent call. The picker owns
its rows, its open/draft state, `CollectionWriteError` and every message;
`useCourseCard` composes it and drops from 260 lines to 173.

The third P1 was not about size — a hook of any length is bound to a list
position when a parent calls it in a map — and `CourseCardItem` already
fixes that by construction.

No behaviour change: the hook's spec is untouched and still passes, which
is what makes the move checkable.

Also trims the barrel to what the three consuming pages actually need. It
was exporting the card's mapper, its formatting helpers and the picker's
tRPC hooks, none of which has a cross-feature caller now that
`CourseCardItem` exists — and each export is a promise to three pages.
