+++ apps/web/features/courses/index.ts
@@ -1,23 +1,33 @@
-export { useCollectionMutations, useMarkCourseTaken } from "./api/mutations";
-export {
-  useCollections,
-  useCourseDetails,
-  useCourseSummaries,
-  useTakenCourses,
-} from "./api/queries";
+/**
+ * The cross-feature API of the courses feature.
+ *
+ * What Explore (#89), Saved (#90) and Collections (#91) need in order to render
+ * a course card, and nothing else. The card's mapper, its formatting helpers
+ * and the tRPC hooks behind the picker are all reachable through
+ * `CourseCardItem`, so they stay inside the feature — a barrel that exports its
+ * own internals turns each of them into a promise to three other pages.
+ */
+
+export { useCourseDetails, useCourseSummaries } from "./api/queries";
+/** The presentational card: takes a model, and measures nothing. */
 export { CourseCard } from "./components/course-card";
-// The binding a screen should render. `useCourseCard` is deliberately not
-// exported: it holds several hooks, so it has to be called from a component
-// that renders one card, and this is that component.
+/**
+ * What a screen renders, keyed by course code.
+ *
+ * `useCourseCard` is deliberately not exported. It holds several hooks, so it
+ * has to be called from a component that renders exactly one card, and this is
+ * that component.
+ */
 export { CourseCardItem } from "./components/course-card-item";
 export { CourseCardWithCharts } from "./components/course-card-with-charts";
 export { CourseDetailsSidebar } from "./components/course-details-sidebar";
 export { CourseItemSkeleton } from "./components/course-item-skeleton";
 export { CoursePageSkeleton } from "./components/course-page-skeleton";
-export type {
-  CourseCardProps,
-  UseCourseCardOptions,
-} from "./hooks/use-course-card";
+export type { UseCourseCardOptions } from "./hooks/use-course-card";
+/**
+ * The collapse ramp. The parent owns it: Explore interpolates from its results
+ * column, Saved and Collections pin an end.
+ */
 export {
   CARD_RAMP_CEILING,
   CARD_RAMP_FLOOR,
@@ -25,11 +35,4 @@ export {
   courseCardGeometry,
   EXPANDED_CARD_GEOMETRY,
 } from "./lib/card-geometry";
-export {
-  type CourseCardCourse,
-  type CourseCardView,
-  formatCount,
-  formatCourseMeta,
-  keywordChips,
-  toCourseCardModel,
-} from "./lib/course-card-model";
+export type { CourseCardCourse } from "./lib/course-card-model";
